0,N Sigmoid Scaler

sigmoid Scales the input to a range of 0,N using the sigmoid function

sigmoid(input, family = "logistic", center = mean, ...)

Arguments

input

(numeric vector or coercible vector)

family

(string | "logistic") The family of sigmoid equation to use: Currently, only "logistic", "generalized", and "gompertz" are supported

center

A function that returns the 'center' of input

...

Valid elements that populate the arguments of type based on the selected family

Value

A numeric vector of domain 0, N

Details

The input must contain positive and negative values when centered == FALSE

Function arguments for each sigmoid family are as follows:

"generalized": list(A, K, C, Q, B, v)

"logistic": list(L, K)

"gompertz": list(A, B, C)

Families tanh, atan, and guder(mannian) only take the input as its argument

References

Generalized Logistic Function

Sigmoid Function

Gompertz Function

Gudermannian Function