Create Bins From Integer Factor

bin.windows creates binned ranges based on the minimum factor of the integer input (i) or user-supplied value.

bin.windows(
  i = 1,
  use.bin = NULL,
  as.factor = FALSE,
  label_format = "<%s,%s>",
  silently = FALSE,
  ...
)

Arguments

i

(integer[]) An integer scalar, vector, or n-dimensional object executed conditionally as follows:

if a vector of length = 1, a zero-based sequence up to abs(i) is used

if a vector of length = 2, a sequence is created from the values in the order given if the input is non-dimensional

if a vector of length >= 3 or i is dimensional, the raw values coerced into a vector

use.bin

The bin size to use: when empty, the smallest prime number or integer factor in i within the range of i is used.

as.factor

(logical) Should the output be converted into a factor?

label_format

(string) A two-argument string compatible with sprintf that controls label output

silently

(logical) Should the output return invisibly?

...

(not used)

Value

A character (or factor) vector the length of the input, as "binned" representations. If the input is dimensional, an array of the same dimensions is returned

Note

Factor output is only available for heterogeneous data structures

See also