Logical Occurrence Map

logic_map conducts an existence test of a vector or list of tuples (fvec) against a vector of unique values (bvec) optionally mapping a supplied vector of values (avec) to all TRUE results.

logic_map(
  fvec,
  avec = 1L,
  bvec = NULL,
  logical.out = FALSE,
  regex = FALSE,
  chatty = FALSE,
  ...
)

Arguments

fvec

(vector) Values to be tested: may be a simple vector or a list of tuples

avec

(vector) Optional vector of numeric values to project across the result (must be the same length as fvec or length-1)

bvec

(vector) A vector of unique values forming the basis of comparison If given as a named list, the output will preserve the names when creating columns; otherwise, the values of bvec are used as the names.

logical.out

(logical, vector) When TRUE, the output consists of logical values; when a vector of length two (2) is supplied, the first value returns on FALSE, the second on TRUE; otherwise, the values of avec are used. If the vector form is used, only the first value at each position is used to supply the choices.

regex

(logical | FALSE) When TRUE, argument bvec is interpreted as patterns against which case-sensitive matches are sought are attempted. This forces the value of test to invoke like.

...

See stri_detect_regex: arguments str and pattern are internally passed and not needed.

Value

An array, the column names being the values of bvec or names of bvec if they exist.

Note

length(fvec) == length(avec)

fvec and bvec must be of compatible types for comparison.

Warning

When combining with a source object, fvec must NOT be sorted during the function call or the values will not map correctly in the output.