Logical Test Occurrence Map

logic_map conducts a test of a vector or list of tuples against a vector of unique values. The test can be one of identity, pattern-matching, or some custom function with Boolean output. Parallelism is supported with a registered foreach backend. If no backend is registered, registerDoSEQ is used as a default.

logic_map(
  fvec,
  avec = NULL,
  bvec = NULL,
  logical.out = FALSE,
  regex = FALSE,
  chatty = FALSE
)

Arguments

fvec

(vector) Values to be tested: may be a simple vector or a list of tuples

avec

(vector) Optional vector of numeric values to project across the result (must be the same length as fvec or length-1)

bvec

(vector) A vector of unique values forming the basis of comparison If given as a named list, the output will preserve the names when creating columns; otherwise, the values of bvec are used as the names.

logical.out

(logical, vector) When TRUE, the output consists of logical values; when a vector of length two (2) is supplied, the first value returns on FALSE, the second on TRUE; otherwise, the values of avec are used. If the vector form is used, only the first value at each position is used to supply the choices.

regex

(logical | FALSE) When TRUE, argument bvec is interpreted as patterns against which case-sensitive matches are sought are attempted. This forces the value of test to invoke like.

chatty

(logical | FALSE) When TRUE, additional information is printed to console

Value

Invisibly, a data.table object, the column names being the values of bvec or names of bvec if they exist

Note

length(fvec) == length(avec)

Warning

When combining with the source object, fvec must NOT be sorted during the function call or the values will not map correctly in the output.