Quantiles Transformation

make.quantiles is a wrapper for quantile replacing the input with calculated values. Unmatched quantiles will return NA in order to match the length of the input vector.

make.quantiles(x, as.factor = FALSE, ...)

Arguments

x

The input vector

as.factor

(logical) Should the output be returned as a factor?

...

(dots_list): Additional arguments sent to quantile. Defaults are as follows:

probs = seq(0, 1, 0.25)

na.rm = FALSE

names = TRUE

type = 7

digits = 7

...

Value

A quantile representation of the input

Note

Any indeterminate probabilities relative to the input will return NA

Examples

sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = TRUE)))()
#>      x    q
#> 1  127  20%
#> 2  705  80%
#> 3  689  80%
#> 4  724  90%
#> 5  346  60%
#> 6  658  70%
#> 7  182  30%
#> 8  124  20%
#> 9  851 100%
#> 10 226  30%
#> 11 726  90%
#> 12 238  50%
#> 13 237  40%
#> 14  20   0%
#> 15 320  50%
#> 16 545  70%
#> 17 230  40%
#> 18 386  60%
#> 19 880 100%
#> 20  43  10%
sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = FALSE)))()
#>      x     q
#> 1  398 445.5
#> 2  588 601.6
#> 3  384 392.4
#> 4  656 683.1
#> 5  557 601.6
#> 6  814 814.0
#> 7  338 392.4
#> 8   77 226.8
#> 9  532 538.8
#> 10 549 554.9
#> 11 678 683.1
#> 12 241 315.2
#> 13 554 554.9
#> 14 262 315.2
#> 15 454 538.8
#> 16  63  63.0
#> 17 437 445.5
#> 18 170 226.8
#> 19  75  76.8
#> 20 729 814.0