Quantiles Transformation

make.quantiles is a wrapper for quantile replacing the input with calculated values. Unmatched quantiles will return NA in order to match the length of the input vector.

make.quantiles(x, ..., as.factor = FALSE)

Arguments

x

The input vector

...

(dots_list): Additional arguments sent to quantile. Defaults are as follows:

probs = seq(0, 1, 0.25)

na.rm = FALSE

names = TRUE

type = 7

digits = 7

...

as.factor

(logical) Should the output be returned as a factor?

Value

A quantile representation of the input

Note

Any indeterminate probabilities relative to the input will return NA

Examples

sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = TRUE)))()
#>      x    q
#> 1  713  90%
#> 2   22  20%
#> 3  400  60%
#> 4  791 100%
#> 5  627  80%
#> 6  767  90%
#> 7   14   0%
#> 8  412  60%
#> 9   24  30%
#> 10 213  50%
#> 11 446  70%
#> 12  34  30%
#> 13 348  50%
#> 14 608  70%
#> 15  16  10%
#> 16  20  20%
#> 17 668  80%
#> 18 123  40%
#> 19 126  40%
#> 20 844 100%
sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = FALSE)))()
#>      x     q
#> 1  696 705.2
#> 2  213 228.4
#> 3   76 164.0
#> 4  811 812.0
#> 5  665 705.2
#> 6  550 581.6
#> 7  136 164.0
#> 8  240 256.2
#> 9  896 896.0
#> 10 631 641.2
#> 11 821 896.0
#> 12 629 641.2
#> 13 742 812.0
#> 14 235 256.2
#> 15  31  31.0
#> 16  38  72.2
#> 17 171 228.4
#> 18 399 581.6
#> 19 332 365.5
#> 20 267 365.5